Update and rename Rollingsales.md to Rollingsales_data_analysis.md
All,

Here are some more changes to the "Rollingsales.md" file in which a new file name is proposed: "Rollingsales_data_analysis.md". This R code and comments has also been modified to represent only the sections containing the "exploratory" analysis of the clean data set. This should be included in the "Analysis Directory".